chore: move expr functions to sub-module
