+++ internal/expressions/function/functions_test.go
@@ -1,4 +1,4 @@
-package expressions
+package function
 
 import (
 	"context"
